requirements.txt
Django==3.1
django-heroku==0.3.1
djangorestframework==3.11.0
django-cors-headers==3.4.0
jsonfield==3.1.0
